continuing last talk works....
Showing new messages in talks as static popups in the talks sidebar.
Added template for a talk item that will be used to refresh the list of
recent talks.
Finally working click "continue talk" link/icon/button.
Simplified VestTalk structure in vest.js
Messages of users are distinguishable.
"Refresh the list of talks in the left sidebar" stub.